Refactor doctor, GPIO channel model, and deploy helpers.
Split doctor into focused modules with a Check builder, centralize
install paths and atomic_write in deploy, and map Channel names to
config GPIO pins without enum discriminants.
